summarizer_WEBBASED/requirements.txt
langchain
google-generativeai
streamlit
tiktoken 
chromadb 
pypdf
